requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
idna==3.3
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.1
urllib3==1.26.8